% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{calculate_jaccard_indices}
\alias{calculate_jaccard_indices}
\title{Calculate Jaccard Indices}
\usage{
calculate_jaccard_indices(networks)
}
\arguments{
\item{networks}{List of named adjacency matrices corresponding to the input networks, same as input to dynetR.}

\item{focus_node}{Character, name of node to focus visualisation on.}
}
\description{
Function to Calculate Jaccard Indices for a List of Networks. Returns a matrix containing the Jaccard indexes for all compared networks in the list.
}
