% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{compare_targeting}
\alias{compare_targeting}
\title{Compare targeting}
\usage{
compare_targeting(input_list)
}
\arguments{
\item{input_list}{List of named adjacency matrices corresponding to the input networks, same as input to dynetR.}
}
\description{
Function to Calculate differential targeting values for all compared networks. Returns a dataframe containing the compared node, networks, targeting values, absoulte difference and log2 ratio of targeting values.
}
