% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{.dynet_verbose}
\alias{.dynet_verbose}
\title{Helper function that prints the results of the subsequent steps}
\usage{
.dynet_verbose(matrix_list)
}
\description{
Helper function that prints the results of the subsequent steps
}
