% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{.edgelist_to_adjacency}
\alias{.edgelist_to_adjacency}
\title{Helper function to format weighted edge list dataframe}
\usage{
.edgelist_to_adjacency(el)
}
\arguments{
\item{el}{An input weighted edge list with from, to, and weight columns}
}
\description{
Formats edge list to adjacency matrix
}
