% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{.expand_adjacency_matrices}
\alias{.expand_adjacency_matrices}
\title{Helper function to expand adjacency matrices to include all unique nodes}
\usage{
.expand_adjacency_matrices(adj_matrices)
}
\arguments{
\item{adj_matrices}{An input list of adjacency matrices}
}
\description{
Expands formatted adjacency matrices to matching sizes
}
