% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{.indata_to_edgelist}
\alias{.indata_to_edgelist}
\title{Helper function to create a combined edgelist out of input adj. matrices}
\usage{
.indata_to_edgelist(ingraph)
}
\arguments{
\item{adj_matrices}{An input  adjacency matrix}
}
\description{
Builds an edgelist out of adj. matrices
}
