% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{.set_rownames_to_colnames}
\alias{.set_rownames_to_colnames}
\title{Helper function to set rownames equal to colnames for a single matrix}
\usage{
.set_rownames_to_colnames(mat)
}
\arguments{
\item{mat}{A single input matrix}
}
\description{
Helps to set matching row and colum nnames
}
