% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{.squareMatrices}
\alias{.squareMatrices}
\title{Helper function to get the square values of the values in the list of matrices}
\usage{
.squareMatrices(matrices)
}
\description{
Helper function to get the square values of the values in the list of matrices
}
