% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{.union_adjacency_matrices}
\alias{.union_adjacency_matrices}
\title{Helper function to create a union of multiple adjacency matrices}
\usage{
.union_adjacency_matrices(inlist)
}
\arguments{
\item{adj_matrices}{An input list of adjacency matrices}
}
\description{
Builds a union adjacency matrix from a list of adjacency matrices
}
