% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{dynetR_plot}
\alias{dynetR_plot}
\title{Visualize the DyNet rewiring results}
\usage{
dynetR_plot(matrix_list, output_dataframe, structure_only = FALSE)
}
\arguments{
\item{matrix_list}{List of adjacency matrices corresponding to the input networks.}

\item{output_dataframe}{Dataframe output from the \code{dynetR} function.}

\item{structure_only}{Logical, should match the \code{structure_only} parameter used in \code{dynetR}. Default is FALSE.}
}
\description{
Generates a network graph visualization based on the output from \code{dynetR}.
}
