% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{format_indata}
\alias{format_indata}
\title{Helper function to format input data, matrix or edge list df}
\usage{
format_indata(input_list)
}
\arguments{
\item{input_list}{A list of matrices or edge lists}
}
\description{
Formats input list data depending on input type
}
