% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{small_multiples_plot}
\alias{small_multiples_plot}
\title{Small multiples plot}
\usage{
small_multiples_plot(input_list, focus_node)
}
\arguments{
\item{focus_node}{Character, name of node to focus visualisation on.}

\item{matrix_list}{List of named adjacency matrices corresponding to the input networks, same as input to dynetR.}
}
\description{
Draws a small multiples plot from the input networks focussing on a specific node
}
