% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetR.R
\name{.centroidDistance}
\alias{.centroidDistance}
\title{Helper function to calculate the centroid distances}
\usage{
.centroidDistance(matrices)
}
\description{
Helper function to calculate the centroid distances
}
